feat: añade para poder eliminar una imagen

+++ src/modules/cattle/application/CattleDelete.ts
@@ -1,9 +1,26 @@
+import { ImageUploader } from "@/modules/images/domain/ImageUploader";
 import { CattleRepository } from "../domain/CattleRepository";
+import { ErrorCattleNotFound } from "../domain/errors/ErrorCattleNotFound";
 
 export class CattleDelete {
-  constructor(private repository: CattleRepository) {}
+  constructor(
+    private repository: CattleRepository,
+    private imgUploader: ImageUploader
+  ) {}
 
   async run(id: number) {
+    const cattle = await this.repository.findById(id);
+
+    if (!cattle) {
+      throw new ErrorCattleNotFound();
+    }
+
+    const image = cattle.getImage();
+
+    if (image) {
+      await this.imgUploader.deleteByUrl(image);
+    }
+
     await this.repository.delete(id);
   }
 }
